__fp16 -> _Float16

+++ util.h
@@ -6,30 +6,30 @@
 
 #define PMASK 0xffffffffffffff
 
-void rand_array(__fp16 *arr, int size) {
+void rand_array(_Float16 *arr, int size) {
   for (int i = 0; i < size; i++)
-    arr[i] = (__fp16)((float)rand() / RAND_MAX - 0.5);
+    arr[i] = (_Float16)((float)rand() / RAND_MAX - 0.5);
 }
 
-void print_mat(__fp16 *arr, int rows, int cols) {
+void print_mat(_Float16 *arr, int rows, int cols) {
   for (int i = 0; i < rows; i++) {
     for (int j = 0; j < cols; j++)
-      printf("%f, ", arr[i*cols+j]);
+      printf("%f, ", (float)arr[i*cols+j]);
     printf("\n");
   }
 }
 
-void read_x(__fp16 ret[256]) {
+void read_x(_Float16 ret[256]) {
   for (uint64_t i = 0; i < 8; i++)
     AMX_STX(PMASK & (uint64_t)(ret+i*32) | (i << 56));
 }
 
-void read_y(__fp16 ret[256]) {
+void read_y(_Float16 ret[256]) {
   for (uint64_t i = 0; i < 8; i++)
     AMX_STY(PMASK & (uint64_t)(ret+i*32) | (i << 56));
 }
 
-void read_z(__fp16 ret[2048]) {
+void read_z(_Float16 ret[2048]) {
   for (uint64_t i = 0; i < 64; i++)
     AMX_STZ(PMASK & (uint64_t)(ret+i*32) | (i << 56));
 }
